add base modal, enable cascade deletion, and styling on ui 🃏

+++ src/server/api/routers/accountDeletionRouter.ts
@@ -1,6 +1,8 @@
 import { PrismaClient } from '@prisma/client';
 import { createTRPCRouter, protectedProcedure } from '../trpc';
 import { z } from 'zod';
+import { prisma } from '~/server/db';
+import bcrypt from 'bcrypt';
 
 const primsa = new PrismaClient();
 
@@ -9,25 +11,67 @@ export const accountDeletionRouter = createTRPCRouter({
 		.input(
 			z.object({
 				userId: z.string(),
+				username: z.string(),
+				password: z.string(),
 			}),
 		)
 		.mutation(async ({ input }) => {
-			// Delete workouts of the user
+			const user = await prisma.user.findUnique({
+				where: {
+					id: input.userId,
+					username: input.username,
+				},
+			});
+
+			if (!user) {
+				throw new Error('User not found');
+			}
+
+			const doesInputPwMatchEncryptedPw = bcrypt.compareSync(
+				input.password as 'string | Buffer',
+				user.password as 'string',
+			);
+
+			if (!doesInputPwMatchEncryptedPw) {
+				throw new Error('Password incorrect');
+			}
+
 			const deletedWorkouts = await primsa.workout.deleteMany({
 				where: {
 					userId: input.userId,
 				},
 			});
 
-			console.log('Number of workouts deleted: ', deletedWorkouts.count);
+			const deletedSessions = await primsa.session.deleteMany({
+				where: {
+					userId: input.userId,
+				},
+			});
 
-			//
+			const deletedRoutines = await primsa.routine.deleteMany({
+				where: {
+					userId: input.userId,
+				},
+			});
 
-			const deletedAccount = await primsa.user.delete({
+			const deletedUser = await primsa.user.delete({
 				where: {
 					id: input.userId,
 				},
 			});
-			return deletedAccount;
+
+			console.log('Number of workouts deleted: ', deletedWorkouts.count);
+			console.log('Number of sessions deleted: ', deletedSessions.count);
+			console.log('Number of routines deleted: ', deletedRoutines.count);
+			console.log('User Deleted: ', deletedUser);
+
+			const allDeletedData = {
+				deletedWorkouts,
+				deletedSessions,
+				deletedRoutines,
+				deletedUser,
+			};
+
+			return allDeletedData;
 		}),
 });
